+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.2 under development
 
-- no changes in this release.
+- Bug #268: Fix foreign keys (@Tigrov)
 
 ## 1.0.1 July 24, 2023